# Rewrite broadcast() and map() based on lift()