--- a/REQUIRE
+++ b/REQUIRE
@@ -1,4 +1,3 @@
-julia 0.4
 julia 0.5
 Compat 0.9.4
 Reexport