--- a/REQUIRE
+++ b/REQUIRE
@@ -1,3 +1,3 @@
 julia 0.5
-Compat 0.9.4
+Compat 0.13.0
 Reexport