@nalimilan 2016-11-13
@davidagold Where do you think `lift` should live? In NullableOps?