@davidagold 2016-11-13
Preferably NullableArrays. Or am I wrong for thinking 'lift' and lifted ops
are sufficiently different strategies that they ought to live in different
packages?

On Sunday, November 13, 2016, Milan Bouchet-Valat <notifications@github.com>
wrote:

> @davidagold <https://github.com/davidagold> Where do you think lift
> should live? In NullableOps?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/JuliaStats/NullableArrays.jl/pull/166#issuecomment-260188991>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ALUCm4fyI9IQ1hlGuon_zLjcVTdMHRJPks5q9xxvgaJpZM4Kv-IA>
> .
>