@nalimilan 2016-11-14
https://github.com/JuliaLang/julia/issues/19313 is kind of annoying: in some cases, the `f.(x)` syntax doesn't work for two-argument functions.

There's another related issue which is even more annoying: loop fusion prevents `lift` from detecting special (blacklisted) functions. So e.g. `identity.(isnull.(x))` returns `NULL` instead of `false` for missing entries.

These two issues mostly affect corner cases, so this change is probably worth it anyway, but in the longer term I'm not sure what to do. The `f.(x)` syntax could be improved in Julia to be more flexible, i.e. calling a method with the function and input types as arguments, which could return a transformed function (lifted, in our case). Then loop fusion would combine these functions instead of the original ones. But I don't see this being accepted in Base without a strong use case, i.e. more or less official support for the lifting framework; we may as well merge https://github.com/JuliaLang/julia/pull/18758 at the same time.

I'm starting to think (again) that all function calls should lift by default; that would be a much more consistent option.