@nalimilan 2016-11-14
So that means lifting by default, without any blacklist?