@bramtayl 2016-11-15
This kind of thing could be integrated pretty easily in ChainMap. Once a lift function is available, I could make

```julia
@chain begin
    x
    @over is.null
    @lift_over identity
end
```

go to

`broadcast(identity, lift(broadcast, is.null, x) )`