@nalimilan 2016-11-25
I've updated the branch to always lift. It now uses a private `broadcast_lift` function, since for now we don't want people to override the lifting behaviour. Surprisingly, new implementation appears to be just as fast as the current one:
```julia
using NullableArrays
X = collect(1:100000);
Y = collect(1:100000);
using BenchmarkTools
@benchmark broadcast!(+, X, X, Y)

# master:
BenchmarkTools.Trial: 
  memory estimate:  112.00 bytes
  allocs estimate:  4
  --------------
  minimum time:     37.605 μs (0.00% GC)
  median time:      38.490 μs (0.00% GC)
  mean time:        42.123 μs (0.00% GC)
  maximum time:     531.757 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%


# This branch:
BenchmarkTools.Trial: 
  memory estimate:  112.00 bytes
  allocs estimate:  4
  --------------
  minimum time:     37.876 μs (0.00% GC)
  median time:      38.707 μs (0.00% GC)
  mean time:        41.929 μs (0.00% GC)
  maximum time:     180.495 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

X = rand(Bool, 100000);
Y = rand(Bool, 100000);
using BenchmarkTools
@benchmark broadcast(&, X, Y)

# master:
BenchmarkTools.Trial: 
  memory estimate:  99.00 kb
  allocs estimate:  38
  --------------
  minimum time:     23.665 μs (0.00% GC)
  median time:      27.348 μs (0.00% GC)
  mean time:        33.014 μs (6.55% GC)
  maximum time:     821.649 μs (90.74% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

# This branch:
BenchmarkTools.Trial: 
  memory estimate:  99.00 kb
  allocs estimate:  38
  --------------
  minimum time:     23.517 μs (0.00% GC)
  median time:      26.476 μs (0.00% GC)
  mean time:        30.080 μs (6.40% GC)
  maximum time:     599.716 μs (92.20% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

Since it passes the tests on 0.5 and 0.6, I think we can seriously consider merging it. It's not yet clear how to get non-standard lifting to work, but fixing standard cases sounds like the highest priority to me.